pyproject.toml
[tool.poetry]
name = "mydomonline"
version = "0.1.0"
description = ""
authors = ["khage"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
envparse = "^0.2.0"
drf-spectacular = "^0.26.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
